Split up prism test data across files
